+++ tests/Controller/Management/GalleryMediaUploadControllerTest.php
@@ -108,8 +108,7 @@ class GalleryMediaUploadControllerTest extends TestCase
         $request = new Request();
         $request->setSession($session);
 
-        $requestStack = new RequestStack();
-        $requestStack->push($request);
+        $requestStack = new RequestStack([$request]);
 
         return [$requestStack, $session];
     }